Undo views + extend model
