+++ decidim-budgets/app/assets/stylesheets/decidim/budgets/_budgets.scss
@@ -1,4 +1,1 @@
 @import "budget/*";
-.hidden{
-  display: none
-}
