+++ decidim-budgets/app/models/decidim/budgets/order.rb
@@ -23,39 +23,14 @@ module Decidim
 
       validates :total_budget, numericality: {
         less_than_or_equal_to: :maximum_budget
-      }, if: :per_budget
-
-      validates :total_project, numericality: {
-        less_than_or_equal_to: :number_of_projects
-      }, if: :per_budget
-
-      validates :total_project, numericality: {
-        less_than_or_equal_to: :number_of_projects
-      }, unless: :per_budget
-
+      }
 
       scope :finished, -> { where.not(checked_out_at: nil) }
       scope :pending, -> { where(checked_out_at: nil) }
 
-      def per_budget
-        !component.settings.vote_per_project?
-      end
-
-      def limit_project_reached?
-        total_project == number_of_projects
-      end
-
       # Public: Returns the sum of project budgets
       def total_budget
         projects.to_a.sum(&:budget)
-      end
-
-      def total_project
-        projects.count
-      end
-
-      def remaining_projects
-        number_of_projects - projects.count
       end
 
       # Public: Returns true if the order has been checked out
@@ -65,11 +40,7 @@ module Decidim
 
       # Public: Check if the order total budget is enough to checkout
       def can_checkout?
-        if component.settings.vote_per_project?
-          limit_project_reached?
-        else
-          total_budget.to_f >= minimum_budget
-        end
+        total_budget.to_f >= minimum_budget
       end
 
       # Public: Returns the order budget percent from the settings total budget
