+++ decidim-budgets/app/views/decidim/budgets/projects/_budget_summary.html.erb
@@ -8,29 +8,22 @@
         </p>
       <% else %>
         <h3 class="heading3"><%= t(".title") %></h3>
-        <p>
-          <% if component_settings.vote_per_project? %>
-            <%= t(".description_per_project", minimum_project: component_settings.total_projects) %>
-          <% else %>
-            <%= t(".description_per_budget", minimum_budget: budget_to_currency(current_order&.minimum_budget)) %>
-          <% end %>
-        </p>
+        <p><%= t(".description", minimum_budget: budget_to_currency(current_order&.minimum_budget)) %></p>
       <% end %>
     <% end %>
-    <% unless component_settings.vote_per_project? %>
-      <div class="budget-summary__total" data-total-budget="<%= component_settings.total_budget %>">
-        <span class="mini-title"><%= t("total_budget") %>
-          <strong class="mini-title__strong mini-title__strong--highlight">
-            <%= budget_to_currency(component_settings.total_budget) %>
-          </strong>
-        </span>
-      </div>
-    <% end %>
+
+    <div class="budget-summary__total" data-total-budget="<%= component_settings.total_budget %>">
+      <span class="mini-title"><%= t("total_budget") %>
+        <strong class="mini-title__strong mini-title__strong--highlight">
+          <%= budget_to_currency(component_settings.total_budget) %>
+        </strong>
+      </span>
+    </div>
 
     <%= render partial: "order_progress" %>
 
     <div>
-      <span class="mini-title <%= 'hidden' if component_settings.vote_per_project? %>">
+      <span class="mini-title">
         <%= t(".assigned") %>
         <%= render partial: "order_total_budget" %>
       </span>
