--- decidim-budgets/app/views/decidim/budgets/projects/_order_total_projects.html.erb
@@ -1,4 +1,0 @@
-<strong id="order-total-projects">
-  <%= component_settings.total_projects %>
-  <%= t(".total_projects", count: component_settings.total_projects) %>
-</strong>
