+++ decidim-budgets/app/views/decidim/budgets/projects/_project.html.erb
@@ -29,7 +29,7 @@
         <span class="card--list__data__number budget-list__number">
           <%= budget_to_currency(project.budget) %>
         </span>
-        <%= action_authorized_button_to "vote", order_line_item_path(project_id: project), method: :post, remote: true, data: { disable: true, budget: project.budget, add: true, "redirect-url": project_path(project) }, disabled: !current_settings.votes_enabled? || current_order_checked_out? || current_order&.limit_project_reached?, class: "button tiny hollow budget--list__action" do %>
+        <%= action_authorized_button_to "vote", order_line_item_path(project_id: project), method: :post, remote: true, data: { disable: true, budget: project.budget, add: true, "redirect-url": project_path(project) }, disabled: !current_settings.votes_enabled? || current_order_checked_out?, class: "button tiny hollow budget--list__action" do %>
           <%= icon("check", class: "icon--small", aria_label: t(".add"), role: "img") %>
         <% end %>
         <% if current_settings.show_votes? %>
