+++ decidim-budgets/config/initializers/extends.rb
@@ -1,1 +1,2 @@
 require "decidim/extends/project_extend.rb"
+require "decidim/extends/order_extend.rb"
